convert to dynamic site
