+++ src/js/main.js
@@ -3,6 +3,9 @@
 // productList(".product-list", "tents");
 
 import productList from "./productList.mjs";
+import { loadHeaderFooter } from "./utils.mjs";
+
+loadHeaderFooter();
 
 productList("#product-container", "tents-category");
